Migrated ApiRequestBaseExtensions tests.
